move checkaccess test - test/access.js
